requirements.txt
Pillow==9.0.1
werkzeug